Update httpHeaderNameString() to return an ASCIILiteral https://bugs.webkit.org/show_bug.cgi?id=257343
Reviewed by Youenn Fablet.

Update httpHeaderNameString() to return an ASCIILiteral instead of a StringView.
It makes it clearer that it only returns ASCII characters and that they are
"immortal".

* Source/WTF/wtf/text/StringHash.h:
(WTF::HashTranslatorASCIILiteral::hash):
(WTF::HashTranslatorASCIILiteral::equal):
(WTF::HashTranslatorASCIILiteral::translate):
(WTF::HashTranslatorASCIILiteralCaseInsensitive::hash):
(WTF::HashTranslatorASCIILiteralCaseInsensitive::equal):
* Source/WTF/wtf/text/WTFString.h:
(WTF::HashTranslatorASCIILiteral::hash): Deleted.
(WTF::HashTranslatorASCIILiteral::equal): Deleted.
(WTF::HashTranslatorASCIILiteral::translate): Deleted.
* Source/WebCore/platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.cpp:
(WebCore::MediaPlayerPrivateGStreamer::handleMessage):
* Source/WebCore/platform/network/HTTPHeaderMap.h:
(WebCore::HTTPHeaderMap::HTTPHeaderMapConstIterator::updateKeyValue):
* Source/WebCore/platform/network/HTTPParsers.cpp:
(WebCore::isCrossOriginSafeHeader):
* Source/WebCore/platform/network/ResourceResponseBase.cpp:
(WebCore::ResourceResponseBase::sanitizeHTTPHeaderFieldsAccordingToTainting):
* Source/WebCore/platform/network/create-http-header-name-table:

Canonical link: https://commits.webkit.org/264587@main
